Add logic to play five rounds and output who won.
There's a bug where the user can't see the output of individual
rounds in the console until the entire game is over -- it seems
like having an active dialog prevents the console from updating?

Since I know this project is going to be made more interactive in
a future version, I'm not going to worry about this for now.