End game after five rounds, add endgame logic
I tried to prepend the final state text element to the current
score element, but couldn't get it to work. StackOverflow points
to that being an issue where vanilla JS nodes are expected, but
the nodes are in fact jQuery nodes. Figure I'll get to that later,
so I just added a blank element in the HTML to fill when needed.